requirements.txt
websocket-client==0.58.0
beautifulsoup4==4.9.3
